prokect_flask/requirements.txt
category-encoders==2.5.1.post0
certifi==2021.10.8
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
patsy==0.5.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
Werkzeug==2.2.2
zipp==3.10.0
